chore: rename `ssa_refactor` module to `ssa`
