+++ .gitignore
@@ -22,5 +22,3 @@ result
 **/target
 !crates/nargo_cli/tests/test_data/*/target
 !crates/nargo_cli/tests/test_data/*/target/witness.tr
-!crates/nargo_cli/tests/test_data_ssa_refactor/*/target
-!crates/nargo_cli/tests/test_data_ssa_refactor/*/target/witness.tr
