+++ crates/noirc_evaluator/src/lib.rs
@@ -7,8 +7,8 @@ mod errors;
 
 // SSA code to create the SSA based IR
 // for functions and execute different optimizations.
-pub mod ssa_refactor;
+pub mod ssa;
 
 pub mod brillig;
 
-pub use ssa_refactor::create_circuit;
+pub use ssa::create_circuit;
