+++ crates/noirc_evaluator/src/ssa/opt/mod.rs
