+++ crates/noirc_evaluator/src/ssa/ssa_gen/context.rs
@@ -9,14 +9,14 @@ use noirc_frontend::monomorphization::ast::{self, LocalId, Parameters};
 use noirc_frontend::monomorphization::ast::{FuncId, Program};
 use noirc_frontend::{BinaryOpKind, Signedness};
 
-use crate::ssa_refactor::ir::dfg::DataFlowGraph;
-use crate::ssa_refactor::ir::function::FunctionId as IrFunctionId;
-use crate::ssa_refactor::ir::function::{Function, RuntimeType};
-use crate::ssa_refactor::ir::instruction::{BinaryOp, Endian, Intrinsic};
-use crate::ssa_refactor::ir::map::AtomicCounter;
-use crate::ssa_refactor::ir::types::{NumericType, Type};
-use crate::ssa_refactor::ir::value::ValueId;
-use crate::ssa_refactor::ssa_builder::FunctionBuilder;
+use crate::ssa::ir::dfg::DataFlowGraph;
+use crate::ssa::ir::function::FunctionId as IrFunctionId;
+use crate::ssa::ir::function::{Function, RuntimeType};
+use crate::ssa::ir::instruction::{BinaryOp, Endian, Intrinsic};
+use crate::ssa::ir::map::AtomicCounter;
+use crate::ssa::ir::types::{NumericType, Type};
+use crate::ssa::ir::value::ValueId;
+use crate::ssa::ssa_builder::FunctionBuilder;
 
 use super::value::{Tree, Value, Values};
